refactor: MovingStrategy 책임 분리
- 인자를 없애고 각 구현체에서 결정할 수 있게 수정
