--- src/main/java/racingcar/strategy/DefaultMovingStrategy.java
@@ -1,11 +1,0 @@
-package racingcar.strategy;
-
-public class DefaultMovingStrategy implements MovingStrategy {
-
-        private static final int MOVING_CONDITION = 4;
-
-        @Override
-        public boolean isMovable(int input) {
-            return input >= MOVING_CONDITION;
-        }
-}
