+++ src/main/java/racingcar/strategy/MovingStrategy.java
@@ -1,5 +1,6 @@
 package racingcar.strategy;
 
+@FunctionalInterface
 public interface MovingStrategy {
-    boolean isMovable(int input);
+    boolean isMovable();
 }
